requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
cffi==1.17.1
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.5.0
idna==3.10
multidict==6.4.2
propcache==0.3.1
psutil==5.9.8
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
yarl==1.19.0
